backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
beautifulsoup4==4.12.3
supabase==2.3.4
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
apscheduler==3.10.4
slowapi==0.1.9
python-jose[cryptography]==3.3.0
dnspython==2.4.2
tldextract==5.1.1
lxml==5.1.0
